- Optimized the VGA rendering process a bit. - main.c: Fixed quitting hanging. - Disabled equalizer on the PSP for a little speedup. - MIDI Synthesizer: Removed most memory protection schemes for a significant speedup (only 1 protection during rendering and voice allocation). -- For some reason the sound is a bit off during certain parts while testing (Speedup according to note number error?) - VGA Blanking is now fully working! - Optimized VGA character generator. - Removed memory protection for the most of the SF2 module (except critical checkpoints).

+++ support/locks.c
@@ -69,7 +69,7 @@ SDL_sem *getLock(char *name)
 	return NULL; //Unable to allocate: invalid lock!
 }
 
-int lock(char *name)
+byte lock(char *name)
 {
 	SDL_sem *lock;
 	lock = getLock(name); //Get the lock!
